backend/requirements.txt
fastapi
uvicorn
python-multipart
gtts
httpx
python-docx
pillow
pytesseract
PyMuPDF
